+++ solution/include/ema-search-int/tools.h
@@ -1,0 +1,6 @@
+#ifndef EMA_SEARCH_TOOLS_H
+#define EMA_SEARCH_TOOLS_H
+
+size_t parse_size(const char *size_str);
+
+#endif
