+++ solution/src/main.c
@@ -12,17 +12,8 @@
 #include "../include/shell/shell.h"
 #include "../include/shell/tools.h"
 #include "../include/ema-search-int/array_generator.h"
+#include "../include/ema-search-int/tools.h"
 
-size_t parse_size(const char *size_str) {
-    char *endptr;
-    errno = 0;
-    size_t size = strtoull(size_str, &endptr, 10);
-    if (errno != 0 || *endptr != '\0') {
-        fprintf(stderr, "Invalid size argument: %s\n", size_str);
-        exit(EXIT_FAILURE);
-    }
-    return size;
-}
 
 int main(int argc, char **argv) {
     char buffer[BUFFER_SIZE];
